♻️ Refactor lib.rs to extract commands and services into modules
Moved command handlers and service logic from lib.rs into organized
module structure:
- commands/recording.rs, diagnostics.rs for Tauri commands
- services/recording.rs, diagnostics.rs for business logic
This improves code organization and maintainability.
