+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,6 @@
+// Group Tauri command implementations by responsibility
+pub mod diagnostics;
+pub use diagnostics::*;
+pub mod recording;
+pub use recording::*;
+
