+++ src-tauri/src/services/mod.rs
@@ -1,0 +1,7 @@
+pub mod diagnostics;
+
+pub use diagnostics::*;
+
+pub mod recording;
+pub use recording::*;
+
